refactor(report): 코드리뷰 반영 (#23)
