+++ src/assets/icons/colored/index.tsx
@@ -1,5 +1,7 @@
+export { default as IcArrowDown } from "./IcArrowDown";
 export { default as IcArrowLeft } from "./IcArrowLeft";
 export { default as IcArrowRight } from "./IcArrowRight";
+export { default as IcArrowUp } from "./IcArrowUp";
 export { default as IcCheckBox } from "./IcCheckBox";
 export { default as IcCheckBoxDefault } from "./IcCheckBoxDefault";
 export { default as IcCopy } from "./IcCopy";
