+++ src/assets/icons/mono/index.tsx
@@ -1,3 +1,1 @@
-export { default as IcArrowDown } from "./IcArrowDown";
-export { default as IcArrowUp } from "./IcArrowUp";
 export { default as IcCalendar } from "./IcCalendar";
